Send every theme change to the system to inform about the user's choice
Also move HedvigApp into its own file to simplify LoggedInActivity
